Add files via upload
The Shubhangi Share Market Online Classes project is designed to offer comprehensive online courses for individuals interested in learning about the share market. The project includes a fully functional website with an integrated registration form and various features to enhance the user experience.

Features -
User Registration: A user-friendly registration form for students to enroll in online classes.
Responsive Design: The website is designed to be responsive, ensuring it works well on various devices.
Interactive UI: An attractive and interactive user interface to engage users effectively.
Terms and Conditions: Clear terms and conditions for the students enrolling in the classes.
Technologies Used
HTML/CSS: For structuring and styling the web pages.